requirements.txt
streamlit>=1.31
openai>=1.40.0
pillow>=10.0.0
python-dotenv>=1.0.1
